fix handler
Committed-by: xiaolei.zl from Dev container
