+++ flex/engines/http_server/service/hqps_service.h
@@ -21,7 +21,7 @@
 #include "flex/engines/graph_db/database/graph_db.h"
 #include "flex/engines/http_server/actor_system.h"
 #include "flex/engines/http_server/handler/admin_http_handler.h"
-#include "flex/engines/http_server/handler/hqps_http_handler.h"
+#include "flex/engines/http_server/handler/graph_db_http_handler.h"
 #include "flex/engines/http_server/workdir_manipulator.h"
 #include "flex/storages/metadata/graph_meta_store.h"
 #include "flex/storages/metadata/metadata_store_factory.h"
@@ -124,7 +124,7 @@ class HQPSService {
  private:
   std::unique_ptr<actor_system> actor_sys_;
   std::unique_ptr<admin_http_handler> admin_hdl_;
-  std::unique_ptr<hqps_http_handler> query_hdl_;
+  std::unique_ptr<graph_db_http_handler> query_hdl_;
   std::atomic<bool> running_{false};
   std::atomic<bool> initialized_{false};
   std::atomic<uint64_t> start_time_{0};
